requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
keras==2.15.0
matplotlib==3.8.3
numpy==1.26.4
pillow==10.2.0
tensorflow==2.15.0
